requirements.txt
fastapi
uvicorn[standard]

faiss-cpu
numpy
requests
python-dotenv

gradio

tensorflow==2.18.0
protobuf==3.20.3

sentence-transformers==2.2.2
